Switch from Explicit Euler integration to Symplectic Euler
Explicit Euler tends to add energy to the system,
which in the worst case can lead to divergence and explosion.

Symplectic Euler is energy conserving.

A simple test is ball bouncing on a fixed plane, both with restitution=1.
With the old euler integration, the ball would quickly bounce higher and higher.
With the new symplectic euler, the ball slowly loses energy instead,
which is much better.

(This energy loss is because gravity is applied before the velocity iterations.
In a later PR I will instead make accelerations from gravity and other external forces
be part of the velocity solver. This should eliminate this energy leakage.)